requirement.txt
alembic==1.18.2
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.31.0
bcrypt==5.0.0
billiard==4.2.4
blis==1.3.3
catalogue==2.0.10
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
cryptography==46.0.4
cymem==2.0.13
Deprecated==1.3.1
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.128.0
filelock==3.20.3
filetype==1.2.0
fsspec==2026.1.0
google-ai-generativelanguage==0.6.15
google-api-core
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-genai==1.68.0
googleapis-common-protos==1.73.0
greenlet==3.3.1
groq==1.1.1
grpcio==1.78.0
grpcio-status
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.3.5
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.3
kombu==5.6.2
limits==5.6.0
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.3
mpmath==1.3.0
murmurhash==1.0.15
networkx==3.6.1
nltk==3.9.2
numpy==2.4.1
packaging==26.0
pandas==3.0.0
pdfminer.six==20251230
pdfplumber==0.11.9
pgvector==0.4.2
pillow==12.1.0
pluggy==1.6.0
preshed==3.0.12
prompt_toolkit==3.0.52
proto-plus==1.27.1
protobuf==5.29.6
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
pypdf==6.6.2
pypdfium2==5.3.0
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-jose==3.5.0
python-magic==0.4.27
python-multipart==0.0.22
PyYAML==6.0.3
RapidFuzz==3.14.3
redis==7.1.0
regex==2026.1.15
requests==2.32.5
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.2
setuptools==80.10.2
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
smart_open==7.5.0
sniffio==1.3.1
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.46
srsly==2.5.2
starlette==0.50.0
structlog==25.5.0
sympy==1.14.0
tenacity==9.1.4
thinc==8.3.10
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.1
transformers==5.0.0
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.40.0
vine==5.1.0
wasabi==1.1.3
watchfiles==1.1.1
wcwidth==0.5.0
weasel==0.4.3
websockets==16.0
wrapt==2.0.1